Merge Cart Done!
Merge Cart is now fully functional and passes all my tests! Also I was able to fix the parsing of the description when reading it from the JSON from the DB, and read in all the special instructions too which are displaying properly. With that - everything about the cart is done. Also - I made the change Josh suggested in Constants.java (which should fix the photo issues), though I have not edited the gitignore to pick up on it.