Now the OrderPage works:
it updates the DB appropriately, and change the user information that is changed like delivery address and phone number, then empties the cart. The order is also showing up in the OrderHistory Page correctly. To do this I had to add phone number to the sql file and respective java files. Additionally, I updated the bundle files with english, and then a placeholder that has Dhoha's name for the french so she can translate. The place order page is super ugly though, so my next task is to make a better UI - but it is functional!