Txt Msg Buggy But Done
I finished the text message service in Place order. It sends a message when the box is checked, and does not when the box is not checked. This also changes the status of the order. However, when I try to place a second order, I get a perpersist error. If I log out and log back in though, I can make another order. Also, using the timer to call the text message service is not working because this requires a new thread which means we cannot rely on our beans that are actively scoped.