Add implementation of Payment Methods
I updated the SQL file because I realized we had a few mistakes in
how long certain fields needed to be. I also increased the number
of possible credit card types to be supported.

I implemented the bean, facade, controller, and validator. I then
built the gui that lists the Payment Methods with an accordion
panel. The forms will call validators as necessary. For the edit
and delete buttons, I had to create action listeners that prepared
a 'selected' object by passing the active tab's payment method
statically when the button got created.

Other minor changes:
* I updated the bundle file.
* I updated the profile page to link to My Payment Methods
* It works now?? Hopefully.